+++ app/src/main/java/com/enpassio/androidmvpandmvvmpatterns/mvvm_pattern/data/ArticleBoundaryCallback.java
@@ -3,19 +3,33 @@ package com.enpassio.androidmvpandmvvmpatterns.mvvm_pattern.data;
 import android.arch.lifecycle.LiveData;
 import android.arch.lifecycle.MutableLiveData;
 import android.arch.paging.PagedList;
+import android.util.Log;
 
+import com.enpassio.androidmvpandmvvmpatterns.BuildConfig;
 import com.enpassio.androidmvpandmvvmpatterns.mvvm_pattern.data.model.Article;
+import com.enpassio.androidmvpandmvvmpatterns.mvvm_pattern.data.model.NewsResponse;
 import com.enpassio.androidmvpandmvvmpatterns.mvvm_pattern.data.network.LocalCache;
 import com.enpassio.androidmvpandmvvmpatterns.mvvm_pattern.data.network.NewsApiService;
+
+import java.util.ArrayList;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 /**
  * Created by Greta Grigutė on 2018-11-18.
  */
 public class ArticleBoundaryCallback extends PagedList.BoundaryCallback<Article> {
-    private static final int NETWORK_PAGE_SIZE = 40;
-    String query;
-    NewsApiService service;
-    LocalCache cache;
+
+    private static final String LOG_TAG = "my_tag";
+    private String query;
+    private NewsApiService service;
+    private LocalCache cache;
+    Boolean insertFinished = false;
+
+    // avoid triggering multiple requests in the same time
+    private Boolean isRequestInProgress = false;
 
     public ArticleBoundaryCallback(String query, NewsApiService service, LocalCache cache) {
         this.query = query;
@@ -23,17 +37,9 @@ public class ArticleBoundaryCallback extends PagedList.BoundaryCallback<Article>
         this.cache = cache;
     }
 
-    // keep the last requested page.
-    // When the request is successful, increment the page number.
-    Integer lastRequestedPage = 1;
-
-    MutableLiveData<String> _networkErrors = new MutableLiveData<>();
-    // LiveData of network errors.
-    LiveData<String> networkErrors;
 
 
-    // avoid triggering multiple requests in the same time
-    Boolean isRequestInProgress = false;
+
 
     @Override
    public void onZeroItemsLoaded() {
@@ -46,22 +52,32 @@ public class ArticleBoundaryCallback extends PagedList.BoundaryCallback<Article>
     }
 
     private void requestAndSaveData(String query) {
-        if (isRequestInProgress) return;
+        if (isRequestInProgress) {return;}
 
                 isRequestInProgress = true;
-//        searchArticles(service, query, lastRequestedPage, NETWORK_PAGE_SIZE, {repos ->
-//                cache.insert(repos, {
-//                        lastRequestedPage++
-//                        isRequestInProgress = false
-//                })
-//        }, {error ->
-//                _networkErrors.postValue(error)
-//                isRequestInProgress = false
-//        })
+        service.getNewsArticles(BuildConfig.NEWS_API_DOT_ORG_KEY,
+                query).enqueue(new Callback<NewsResponse>() {
+            @Override
+            public void onResponse(Call<NewsResponse> call, Response<NewsResponse> response) {
+                if (response.isSuccessful()) {
+                    ArrayList<Article> responseResults = (ArrayList<Article>) response.body().getArticles();
+                    Log.d(LOG_TAG, "Getting the reponse size: " + responseResults.size());
+
+                   insertFinished = cache.insert(responseResults, insertFinished);
+                     isRequestInProgress = false;
+                } else {
+                    Log.d(LOG_TAG, "Getting Error");
+                    isRequestInProgress = false;
+                }
+            }
+
+            @Override
+            public void onFailure(Call<NewsResponse> call, Throwable t) {
+                Log.e(LOG_TAG, "error is: " + t.getMessage());
+                isRequestInProgress = false;
+            }
+
+        });
     }
 
-
-    MutableLiveData<String> get() {
-        return _networkErrors;
-    }
 }
